Luybo

+++ Concert/concertApp/src/main/java/com/concert/concertApp/controllers/ConcertHallController.java
@@ -1,6 +1,8 @@
 package com.concert.concertApp.controllers;
 
+import com.concert.concertApp.entities.City;
 import com.concert.concertApp.entities.ConcertHall;
+import com.concert.concertApp.repositories.CityRepository;
 import com.concert.concertApp.repositories.ConcertHallRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,8 +19,11 @@ import java.util.Optional;
 public class ConcertHallController {
 
     private final ConcertHallRepository concertHallRepo;
-    ConcertHallController(ConcertHallRepository concertHallRepo) {
+    private final CityRepository cityRepo ;
+    
+    ConcertHallController(ConcertHallRepository concertHallRepo, CityRepository cityRepo) {
         this.concertHallRepo = concertHallRepo;
+        this.cityRepo = cityRepo;
     }
 
     @GetMapping("/fetch")
@@ -46,12 +51,12 @@ public class ConcertHallController {
         try {
 
             hall = concertHallRepo.findConcertHallByConHallId(id)
-                    .orElse(new ConcertHall(name, adress, city, capacity));
+                    .orElse(new ConcertHall(name, adress,  capacity));
 
             if (hall != null) {
                 hall.setConHallName(name);
                 hall.setConHallAdress(adress);
-                hall.setConHallCity(city);
+               // hall.setConHallCity(city);
                 hall.setConHallCapacity(capacity);
             }
 
@@ -65,16 +70,47 @@ public class ConcertHallController {
     }
     @DeleteMapping("/deleteHall")
     public ResponseEntity<?> deleteHall (String name , String city ){
-        Optional<ConcertHall> hall = concertHallRepo.findConcertHallByConHallCityAndConHallName(city , name);
-        if(hall.isEmpty()){
-            return ResponseEntity.ok("Hall not found");
 
+        Optional<City> city1 = cityRepo.findByName(city);
+        if(!city1.isEmpty()) {
+
+            Optional<ConcertHall> hall = concertHallRepo.findConcertHallByConHallName(name);
+            if (hall.isEmpty()) {
+                return ResponseEntity.ok("Hall not found");
+
+            }
+            concertHallRepo.delete(hall.get());
+            return ResponseEntity.ok("Concert hall with name : " + name + "in City" + city +
+                    "was deleted");
+            //return("Hall with name : " + name + "in " + city + " was deleted!");
+
+        }else
+            return ResponseEntity.ok("City not found");
+    }
+
+    @PostMapping("/city/save")
+    public  ResponseEntity<?> saveCity(String name ,
+                                       @RequestParam(required = false) Long id ){
+        City city  ;
+        try{
+            city = cityRepo.findById(id)
+                    .orElse(new City(name));
+
+            if(city != null)
+            {
+                city.setName(name);
+            }
+
+        }catch (Exception e){
+            return  new ResponseEntity<>(e.getMessage(), HttpStatus.OK);
         }
-concertHallRepo.delete(hall.get()) ;
-        return ResponseEntity.ok( "Concert hall with name : " + name +
-                " in city :" + city + "was deleted" );
-        //return("Hall with name : " + name + "in " + city + " was deleted!");
 
+        cityRepo.save(city);
+        return ResponseEntity.ok("City was save" );
+    }
 
+    @GetMapping("/city/fetch")
+    public List<City> getAllCitites() {
+        return cityRepo.findAll();
     }
 }
